# v2_old/api/cmd/server/api/openapi.yaml
openapi: 3.0.3
info:
  title: Gallformers API
  version: 2.0.0
  description: |
    REST API for gallformers.org - a comprehensive database and reference guide for galls,
    the abnormal plant growths caused by insects, mites, and other organisms.

    ## Authentication

    Most read endpoints are public. Write operations (POST, PUT, DELETE) require authentication
    via Auth0 JWT token stored in an httpOnly cookie.

    ## Rate Limiting

    No rate limiting is currently implemented.
  contact:
    name: Gallformers
    url: https://gallformers.org
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://gallformers.fly.dev
    description: Staging server (Fly.io)
  - url: https://gallformers.org
    description: Production server
  - url: http://localhost:8080
    description: Local development server
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
      description: |
        Auth0 JWT token stored in httpOnly cookie. Set after successful login
        via the /api/v2/auth/callback endpoint.
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: Machine-readable error code
              example: NOT_FOUND
            message:
              type: string
              description: Human-readable error message
              example: Gall with ID 123 not found
          required:
            - code
            - message
      required:
        - error
...
    Alias:
      type: object
      description: An alias (alternative name) for a species
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier
        name:
          type: string
          description: The alias name
        type:
          type: string
          enum: [common, scientific]
          description: Type of alias
        description:
          type: string
          description: Optional description of the alias
      required:
        - id
        - name
        - type
        - description
    FilterField:
      type: object
      description: A filter field value (color, shape, texture, etc.)
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier
        field:
          type: string
          description: The field value
        description:
          type: string
          nullable: true
          description: Optional description
      required:
        - id
        - field
    Host:
      type: object
      description: A host plant for a gall
      properties:
        id:
          type: integer
          format: int64
          description: Species ID of the host
        name:
          type: string
          description: Name of the host species
      required:
        - id
        - name
    Gall:
      type: object
      description: A gall species with all associated data
      properties:
        id:
          type: integer
          format: int64
          description: Species ID
        name:
          type: string
          description: Scientific name
        gall_id:
          type: integer
          format: int64
          description: Gall table ID
        datacomplete:
          type: boolean
          description: Whether data is complete
        abundance_id:
          type: integer
          format: int64
          nullable: true
          description: Abundance level ID
        detachable:
          type: integer
          nullable: true
          description: Detachability value (0=none, 1=integral, 2=detachable, 3=both)
        undescribed:
          type: boolean
          description: Whether this is an undescribed species
        aliases:
          type: array
          items:
            $ref: '#/components/schemas/Alias'
        hosts:
          type: array
          items:
            $ref: '#/components/schemas/Host'
        colors:
          type: array
          items:
            $ref: '#/components/schemas/FilterField'
        shapes:
          type: array
          items:
            $ref: '#/components/schemas/FilterField'
        textures:
          type: array
          items:
            $ref: '#/components/schemas/FilterField'
        locations:
          type: array
          items:
            $ref: '#/components/schemas/FilterField'
        alignments:
          type: array
          items:
            $ref: '#/components/schemas/FilterField'
        walls:
          type: array
          items:
            $ref: '#/components/schemas/FilterField'
        cells:
          type: array
          items:
            $ref: '#/components/schemas/FilterField'
        seasons:
          type: array
          items:
            $ref: '#/components/schemas/FilterField'
        forms:
          type: array
          items:
            $ref: '#/components/schemas/FilterField'
      required:
        - id
        - name
        - gall_id
        - datacomplete
        - undescribed
        - aliases
    GallListResponse:
      type: object
      description: Paginated list of galls
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Gall'
        total:
          type: integer
          format: int64
          description: Total number of galls matching query
        limit:
          type: integer
          nullable: true
          description: Limit applied to results (if any)
        offset:
          type: integer
          description: Offset applied to results
      required:
        - data
        - total
        - offset
    GallCreateRequest:
      type: object
      description: Request body for creating a gall
      properties:
        name:
          type: string
          description: Scientific name (required)
        datacomplete:
          type: boolean
          description: Whether data is complete
        abundance_id:
          type: integer
          format: int64
          nullable: true
        detachable:
          type: integer
          nullable: true
          description: Detachability value
        undescribed:
          type: boolean
        aliases:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              type:
                type: string
                enum: [common, scientific]
              description:
                type: string
        hosts:
          type: array
          items:
            type: integer
            format: int64
          description: Array of host species IDs
        colors:
          type: array
          items:
            type: integer
            format: int64
        shapes:
          type: array
          items:
            type: integer
            format: int64
        textures:
          type: array
          items:
            type: integer
            format: int64
        locations:
          type: array
          items:
            type: integer
            format: int64
        alignments:
          type: array
          items:
            type: integer
            format: int64
        walls:
          type: array
          items:
            type: integer
            format: int64
        cells:
          type: array
          items:
            type: integer
            format: int64
        seasons:
          type: array
          items:
            type: integer
            format: int64
        forms:
          type: array
          items:
            type: integer
            format: int64
      required:
        - name
    GallUpdateRequest:
      type: object
      description: Request body for updating a gall
      properties:
        name:
          type: string
          description: Scientific name (required)
        datacomplete:
          type: boolean
        abundance_id:
          type: integer
          format: int64
          nullable: true
        detachable:
          type: integer
          nullable: true
        undescribed:
          type: boolean
        aliases:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              type:
                type: string
                enum: [common, scientific]
              description:
                type: string
        hosts:
          type: array
          items:
            type: integer
            format: int64
        colors:
          type: array
          items:
            type: integer
            format: int64
        shapes:
          type: array
          items:
            type: integer
            format: int64
        textures:
          type: array
          items:
            type: integer
            format: int64
        locations:
          type: array
          items:
            type: integer
            format: int64
        alignments:
          type: array
          items:
            type: integer
            format: int64
        walls:
          type: array
          items:
            type: integer
            format: int64
        cells:
          type: array
          items:
            type: integer
            format: int64
        seasons:
          type: array
          items:
            type: integer
            format: int64
        forms:
          type: array
          items:
            type: integer
            format: int64
      required:
        - name
    HostResponse:
      type: object
      description: A host plant species with all associated data
      properties:
        id:
          type: integer
          format: int64
          description: Species ID
        name:
          type: string
          description: Scientific name
        datacomplete:
          type: boolean
          description: Whether data is complete
        abundance_id:
          type: integer
          format: int64
          nullable: true
          description: Abundance level ID
        abundance:
          type: string
          nullable: true
          description: Abundance level name
        aliases:
          type: array
          items:
            $ref: '#/components/schemas/Alias'
        places:
          type: array
          items:
            $ref: '#/components/schemas/Place'
        galls:
          type: array
          items:
            $ref: '#/components/schemas/Host'
          description: Galls associated with this host
      required:
        - id
        - name
        - datacomplete
        - aliases
    HostSimpleResponse:
      type: object
      description: A simplified host response for list views
      properties:
        id:
          type: integer
          format: int64
          description: Species ID
        name:
          type: string
          description: Scientific name
        datacomplete:
          type: boolean
          description: Whether data is complete
        aliases:
          type: array
          items:
            $ref: '#/components/schemas/Alias'
        places:
          type: array
          items:
            $ref: '#/components/schemas/Place'
      required:
        - id
        - name
        - datacomplete
        - aliases
        - places
    HostListResponse:
      type: object
      description: Paginated list of hosts
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/HostResponse'
        total:
          type: integer
          format: int64
          description: Total number of hosts matching query
        limit:
          type: integer
          nullable: true
          description: Limit applied to results (if any)
        offset:
          type: integer
          description: Offset applied to results
      required:
        - data
        - total
        - offset
    HostSimpleListResponse:
      type: object
      description: Paginated list of simplified hosts
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/HostSimpleResponse'
        total:
          type: integer
          format: int64
          description: Total number of hosts matching query
        limit:
          type: integer
          nullable: true
          description: Limit applied to results (if any)
        offset:
          type: integer
          description: Offset applied to results
      required:
        - data
        - total
        - offset
    HostCreateRequest:
      type: object
      description: Request body for creating a host
      properties:
        name:
          type: string
          description: Scientific name (required)
        datacomplete:
          type: boolean
          description: Whether data is complete
        abundance_id:
          type: integer
          format: int64
          nullable: true
        aliases:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              type:
                type: string
                enum: [common, scientific]
              description:
                type: string
        places:
          type: array
          items:
            type: integer
            format: int64
          description: Array of place IDs
      required:
        - name
    HostUpdateRequest:
      type: object
      description: Request body for updating a host
      properties:
        name:
          type: string
          description: Scientific name (required)
        datacomplete:
          type: boolean
        abundance_id:
          type: integer
          format: int64
          nullable: true
        aliases:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              type:
                type: string
                enum: [common, scientific]
              description:
                type: string
        places:
          type: array
          items:
            type: integer
            format: int64
          description: Array of place IDs
      required:
        - name
...
    Place:
      type: object
      description: A geographic location (continent, country, state, etc.)
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier
        name:
          type: string
          description: Name of the place
        code:
          type: string
          description: Code/abbreviation for the place
        type:
          type: string
          enum: [continent, country, region, state, province, county, city]
          description: Type of place
      required:
        - id
        - name
        - code
        - type
...
    GallHost:
      type: object
      description: A gall-host relationship (which host plants a gall is found on)
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier for the relationship
        gall_species_id:
          type: integer
          format: int64
          description: Species ID of the gall
        host_species_id:
          type: integer
          format: int64
          description: Species ID of the host plant
        host_name:
          type: string
          description: Name of the host plant species
      required:
        - id
        - gall_species_id
        - host_species_id
    GallHostListResponse:
      type: object
      description: List of gall-host relationships
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/GallHost'
        total:
          type: integer
          format: int64
          description: Total number of host relationships for the gall
      required:
        - data
        - total
    GallHostCreateRequest:
      type: object
      description: Request body for creating a gall-host relationship
      properties:
        gall_species_id:
          type: integer
          format: int64
          description: Species ID of the gall (required)
        host_species_id:
          type: integer
          format: int64
          description: Species ID of the host plant (required)
      required:
        - gall_species_id
        - host_species_id
    GallHostDeleteRequest:
      type: object
      description: Request body for deleting a gall-host relationship
      properties:
        gall_species_id:
          type: integer
          format: int64
          description: Species ID of the gall (required)
        host_species_id:
          type: integer
          format: int64
          description: Species ID of the host plant (required)
      required:
        - gall_species_id
        - host_species_id
...
  parameters:
    limitParam:
      name: limit
      in: query
      description: Maximum number of results to return
      schema:
        type: integer
        minimum: 1
        maximum: 1000
    offsetParam:
      name: offset
      in: query
      description: Number of results to skip
      schema:
        type: integer
        minimum: 0
        default: 0
    searchParam:
      name: q
      in: query
      description: Search term
      schema:
        type: string
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: BAD_REQUEST
              message: Invalid request parameters
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: UNAUTHORIZED
              message: No auth token
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: NOT_FOUND
              message: Resource not found
...
paths:
...
  /api/v2/galls:
    get:
      tags:
        - Galls
      summary: List or search galls
      description: |
        Returns a list of all galls, optionally filtered by search query.
        Supports pagination via limit and offset parameters.
      operationId: listGalls
      parameters:
        - $ref: '#/components/parameters/searchParam'
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/offsetParam'
      responses:
        '200':
          description: List of galls
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GallListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
    post:
      tags:
        - Galls
      summary: Create a new gall
      description: Creates a new gall species entry. Requires authentication.
      operationId: createGall
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GallCreateRequest'
      responses:
        '201':
          description: Gall created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Gall'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v2/galls/{id}:
    get:
      tags:
        - Galls
      summary: Get a gall by ID
      description: Returns detailed information about a specific gall.
      operationId: getGallById
      parameters:
        - name: id
          in: path
          required: true
          description: The species ID of the gall
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Gall details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Gall'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags:
        - Galls
      summary: Update a gall
      description: Updates an existing gall. Requires authentication.
      operationId: updateGall
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The species ID of the gall
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GallUpdateRequest'
      responses:
        '200':
          description: Gall updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Gall'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
        - Galls
      summary: Delete a gall
      description: Deletes a gall and all associated data. Requires authentication.
      operationId: deleteGall
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The species ID of the gall
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Gall deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # Host endpoints
  /api/v2/hosts:
    get:
      tags:
        - Hosts
      summary: List or search hosts
      description: |
        Returns a list of all host plants, optionally filtered by search query.
        Supports pagination via limit and offset parameters.
        Use the `simple` query parameter for a simplified response format.
      operationId: listHosts
      parameters:
        - $ref: '#/components/parameters/searchParam'
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/offsetParam'
        - name: simple
          in: query
          description: If present, returns simplified host data (id, name, aliases, places)
          schema:
            type: boolean
      responses:
        '200':
          description: List of hosts
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/HostListResponse'
                  - $ref: '#/components/schemas/HostSimpleListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
    post:
      tags:
        - Hosts
      summary: Create a new host
      description: Creates a new host plant species entry. Requires authentication.
      operationId: createHost
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HostCreateRequest'
      responses:
        '201':
          description: Host created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HostResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v2/hosts/{id}:
    get:
      tags:
        - Hosts
      summary: Get a host by ID
      description: Returns detailed information about a specific host plant.
      operationId: getHostById
      parameters:
        - name: id
          in: path
          required: true
          description: The species ID of the host
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Host details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HostResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags:
        - Hosts
      summary: Update a host
      description: Updates an existing host plant. Requires authentication.
      operationId: updateHost
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The species ID of the host
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HostUpdateRequest'
      responses:
        '200':
          description: Host updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HostResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
        - Hosts
      summary: Delete a host
      description: Deletes a host plant and all associated data. Requires authentication.
      operationId: deleteHost
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The species ID of the host
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Host deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v2/gall-hosts:
    get:
      tags:
        - GallHosts
      summary: List hosts for a gall
      description: |
        Returns a list of host plants associated with a specific gall.
        The gallid query parameter is required.
      operationId: listGallHosts
      parameters:
        - name: gallid
          in: query
          required: true
          description: The species ID of the gall
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: List of gall-host relationships
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GallHostListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
    post:
      tags:
        - GallHosts
      summary: Create a gall-host relationship
      description: Creates a new association between a gall and a host plant. Requires authentication.
      operationId: createGallHost
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GallHostCreateRequest'
      responses:
        '201':
          description: Gall-host relationship created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GallHost'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
    delete:
      tags:
        - GallHosts
      summary: Delete a gall-host relationship
      description: Deletes an association between a gall and a host plant. Requires authentication.
      operationId: deleteGallHost
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GallHostDeleteRequest'
      responses:
        '204':
          description: Gall-host relationship deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # Species-Source relationship endpoints
...